pyproject.toml
[tool.poetry]
name = "fast-api-madr"
version = "0.1.0"
description = ""
authors = ["oseiasdfarias <oseias.dfarias@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = { extras = ["standard"], version = "^0.112.0" }
psycopg = {extras = ["binary"], version = "^3.2.1"}
pyjwt = "^2.9.0"
python-multipart = "^0.0.9"
pydantic-settings = "^2.4.0"
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.5.7"
httpx = "^0.27.0"
testcontainers = "^4.7.2"
factory-boy = "^3.3.0"


[tool.pytest.init_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.ruff]
line-length = 79
extend-exclude = ["migrations"]

[tool.ruff.lint]
preview = true
select = ["I", "F", "E", "W", "PL", "PT"]

[tool.ruff.lint.pylint]
max-args = 10

[tool.taskipy.tasks]
run = "fastapi dev fast_api_madr/app.py"

pre_test = "task lint"
test = "pytest --cov=fast_api_madr -vv"
post_test = "coverage html"

lint = "ruff check .; ruff check . --diff"
format = "ruff check . --fix; ruff format ."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
